Add Firebase functions and configuration for detective case generation
- Introduced Firebase functions for generating and seeding detective cases, including scheduled case generation.
- Added TypeScript support with ESLint configuration for code quality.
- Implemented Firestore integration for managing detective cases and clues.
- Updated project structure to include new functions and utilities for case handling.
- Enhanced validation for detective cases to ensure data integrity and compliance with game rules.
- Updated `.gitignore` to exclude unnecessary files and directories.
- Added necessary configuration files for TypeScript and Jest testing.
- Updated `pubspec.yaml` and `pubspec.lock` to reflect new dependencies and versioning.
- Created utility functions for date handling and answer word validation.
- Implemented tests for case validation and parsing logic to ensure reliability.

+++ lib/core/utils/game_layout_metrics.dart
@@ -64,7 +64,7 @@ class GameLayoutMetrics {
 
     var tileSize = widthTile < heightTile
         ? widthTile
-        : (heightTile.clamp(_minTileSize, widthTile) as double);
+        : heightTile.clamp(_minTileSize, widthTile);
     var scale = tileSize / widthTile;
 
     if (flags.hasKeyboard) {
@@ -78,7 +78,7 @@ class GameLayoutMetrics {
       heightTile = (gridHeight - 4 * _defaultSpacing) / 5;
       tileSize = widthTile < heightTile
           ? widthTile
-          : (heightTile.clamp(_minTileSize, widthTile) as double);
+          : heightTile.clamp(_minTileSize, widthTile);
       scale = tileSize / widthTile;
     }
